refactor(frontend): organization and workspace APIs move to @agenta/entities
Members, Organizations and Access & Security all need org and workspace data, and
none of it was reachable from a package. The two service modules move to
@agenta/entities/organization, which is what /m has to import from — it cannot
reach @/oss/*.

The org and workspace TYPES had been declared in oss/lib/Types and are now
declared once, in the package. That mattered: two structurally different
`Workspace` types made jotai-tanstack-query give up on inferring the org query,
which showed up as 69 `NonNullable<TQueryFnData>` errors across 19 files that had
nothing to do with the move.

Two reads used the app's fetch client, which a package cannot reach; they now use
the shared axios instance like the rest, keeping their empty-not-throw contract.
`getProjectValues()` becomes a read of `projectIdAtom`, which both apps hydrate.

The `_ignoreError` / `_skipAuthUpgradeRedirect` request flags are now declared on
AxiosRequestConfig in @agenta/shared/api, so the 20 call sites that carried
`as any` to set them no longer need the cast.

Also fixes /m's Projects tab, which was passing a hardcoded empty array and so
could only ever render its empty state.

Gates: lint 24/24, tsc 0 across entities, oss, ee and mobile.
