+++ web/oss/src/components/pages/settings/WorkspaceManage/WorkspaceManage.tsx
@@ -1,5 +1,7 @@
 import {useCallback, useMemo, useState, type FC} from "react"
 
+import type {WorkspaceMember} from "@agenta/entities/organization"
+import {removeFromWorkspace, resendInviteToWorkspace} from "@agenta/entities/organization"
 import {useStaticTable} from "@agenta/settings"
 import {formatDay} from "@agenta/shared/utils/dateTime"
 import {message} from "@agenta/ui/app-message"
@@ -21,9 +23,7 @@ import {useWorkspacePermissions} from "@/oss/hooks/useWorkspacePermissions"
 import {isEE, isEmailInvitationsEnabled} from "@/oss/lib/helpers/isEE"
 import {useEntitlements} from "@/oss/lib/helpers/useEntitlements"
 import {getUsernameFromEmail} from "@/oss/lib/helpers/utils"
-import {WorkspaceMember} from "@/oss/lib/Types"
 import {updateUsername} from "@/oss/services/profile"
-import {removeFromWorkspace, resendInviteToWorkspace} from "@/oss/services/workspace/api"
 import {useOrgData} from "@/oss/state/org"
 import {useProfileData} from "@/oss/state/profile"
 import {useWorkspaceMembers} from "@/oss/state/workspace"
