+++ web/oss/src/lib/atoms/organization.ts
@@ -1,5 +1,4 @@
+import type {WorkspaceRole} from "@agenta/entities/organization"
 import {atom} from "jotai"
-
-import {WorkspaceRole} from "../Types"
 
 export const workspaceRolesAtom = atom<WorkspaceRole[]>([])
