+++ web/oss/src/pages/workspaces/accept.tsx
@@ -1,5 +1,6 @@
 import {useEffect, useRef, useState, type FC} from "react"
 
+import {acceptWorkspaceInvite} from "@agenta/entities/organization"
 import {message} from "@agenta/ui/app-message"
 import {Button, Card, Typography} from "antd"
 import {useAtomValue} from "jotai"
@@ -11,7 +12,6 @@ import ContentSpinner from "@/oss/components/Spinner/ContentSpinner"
 import {inviteErrorMessageFromCode} from "@/oss/lib/helpers/authMessages"
 import {isEE} from "@/oss/lib/helpers/isEE"
 import {getJWT} from "@/oss/services/api"
-import {acceptWorkspaceInvite} from "@/oss/services/workspace/api"
 import {useOrgData} from "@/oss/state/org"
 import {cacheWorkspaceOrgPair} from "@/oss/state/org/selectors/org"
 import {useProjectData} from "@/oss/state/project"
