+++ web/oss/src/state/org/selectors/org.ts
@@ -1,3 +1,5 @@
+import type {Org, OrgDetails} from "@agenta/entities/organization"
+import {fetchAllOrgsList, fetchSingleOrg} from "@agenta/entities/organization"
 import {fetchProject} from "@agenta/entities/project"
 import {idleReadyAtom, logAtom} from "@agenta/shared/state"
 import type {User} from "@agenta/shared/types"
@@ -5,8 +7,6 @@ import {atom} from "jotai"
 import {atomWithQuery} from "jotai-tanstack-query"
 
 import {queryClient} from "@/oss/lib/api/queryClient"
-import {Org, OrgDetails} from "@/oss/lib/Types"
-import {fetchAllOrgsList, fetchSingleOrg} from "@/oss/services/organization/api"
 import {appIdentifiersAtom, appStateSnapshotAtom, requestNavigationAtom} from "@/oss/state/appState"
 import {userAtom} from "@/oss/state/profile/selectors/user"
 import {sessionExistsAtom} from "@/oss/state/session"
