+++ web/packages/agenta-entities/package.json
@@ -47,6 +47,7 @@
         "./loadable": "./src/loadable/index.ts",
         "./profile": "./src/profile/index.ts",
         "./project": "./src/project/index.ts",
+        "./organization": "./src/organization/index.ts",
         "./queue": "./src/queue/index.ts",
         "./runnable": "./src/runnable/index.ts",
         "./runnable/bridge": "./src/runnable/bridge.ts",
