+++ web/packages/agenta-entities/src/organization/api.ts
@@ -1,8 +1,13 @@
-import axios from "@/oss/lib/api/assets/axiosConfig"
-import {getAgentaApiUrl} from "@/oss/lib/helpers/api"
+import {axios, getAgentaApiUrl} from "@agenta/shared/api"
 
-import {fetchJson, getBaseUrl} from "../../../lib/api/assets/fetchClient"
-import {Org, OrgDetails} from "../../../lib/Types"
+import type {
+    Org,
+    OrgDetails,
+    OrganizationDomain,
+    OrganizationProvider,
+    OrganizationProviderSettings,
+    OrganizationUpdatePayload,
+} from "./types"
 
 export const checkOrganizationAccess = async (organizationId: string) => {
     const response = await axios.get(`${getAgentaApiUrl()}/auth/access`, {
@@ -10,7 +15,7 @@ export const checkOrganizationAccess = async (organizationId: string) => {
         _skipAuthUpgradeRedirect: true,
         _ignoreError: true,
         validateStatus: () => true,
-    } as any)
+    })
 
     const detailError = response.data?.detail?.error
     if (response.status >= 200 && response.status < 300 && !detailError) {
@@ -21,94 +26,34 @@ export const checkOrganizationAccess = async (organizationId: string) => {
 }
 
 /**
- * Fetch all organizations using modern fetchJson
- * Replaces the old axios-based fetchAllOrgsList
+ * Both reads go through the shared axios instance, which carries auth via its interceptor — the
+ * desktop split them onto a separate fetch client for no behavioural reason, and a package cannot
+ * reach that client anyway. Each still answers empty rather than throwing, as its callers expect.
  */
 export const fetchAllOrgsList = async (): Promise<Org[]> => {
-    const base = getBaseUrl()
-    const url = new URL("api/organizations", base)
-
-    console.log("🔍 Organizations fetcher debug:", {
-        base,
-        url: url.toString(),
-    })
-
     try {
-        console.log("🚀 Calling fetchJson with URL:", url.toString())
-        const data = await fetchJson(url)
-        console.log("✅ Organizations fetcher success:", {
-            count: data?.length || 0,
-        })
-        return data || []
-    } catch (error: any) {
-        console.error("❌ Organizations fetcher failed:", {
-            message: error?.message,
-            status: error?.status,
-            statusText: error?.statusText,
-            url: url.toString(),
-            stack: error?.stack?.split("\n").slice(0, 3).join("\n"),
-        })
-        // Return empty array instead of throwing to prevent test failures
+        const {data} = await axios.get(`${getAgentaApiUrl()}/organizations`)
+        return Array.isArray(data) ? data : []
+    } catch (error) {
+        if ((error as {response?: {status?: number}})?.response?.status === 401) return []
+        console.error("Failed to fetch organizations", error)
         return []
     }
 }
 
-/**
- * Fetch single organization details using modern fetchJson
- * Replaces the old axios-based fetchSingleOrg
- */
 export const fetchSingleOrg = async ({
     organizationId,
 }: {
     organizationId: string
 }): Promise<OrgDetails | null> => {
-    const base = getBaseUrl()
-    const url = new URL(`api/organizations/${organizationId}`, base)
-
-    console.log("🔍 Single organization fetcher debug:", {
-        base,
-        organizationId,
-        url: url.toString(),
-    })
-
     try {
-        console.log("🚀 Calling fetchJson with URL:", url.toString())
-        const data = await fetchJson(url)
-        console.log("✅ Single organization fetcher success:", {
-            organizationId,
-            name: data?.name,
-        })
-        return data
-    } catch (error: any) {
-        console.error("❌ Single organization fetcher failed:", {
-            message: error?.message,
-            status: error?.status,
-            statusText: error?.statusText,
-            url: url.toString(),
-            stack: error?.stack?.split("\n").slice(0, 3).join("\n"),
-        })
-        // Return null instead of throwing to prevent test failures
+        const {data} = await axios.get(`${getAgentaApiUrl()}/organizations/${organizationId}`)
+        return data ?? null
+    } catch (error) {
+        if ((error as {response?: {status?: number}})?.response?.status === 401) return null
+        console.error("Failed to fetch organization", organizationId, error)
         return null
     }
-}
-
-// Partial flags interface for PATCH updates
-export interface OrganizationFlagsUpdate {
-    is_personal?: boolean
-    is_demo?: boolean
-    allow_email?: boolean
-    allow_social?: boolean
-    allow_sso?: boolean
-    auto_join?: boolean
-    domains_only?: boolean
-    allow_root?: boolean
-}
-
-export interface OrganizationUpdatePayload {
-    slug?: string
-    name?: string
-    description?: string
-    flags?: OrganizationFlagsUpdate
 }
 
 export const updateOrganization = async (
@@ -121,7 +66,7 @@ export const updateOrganization = async (
         payload,
         {
             _ignoreError: ignoreAxiosError,
-        } as any,
+        },
     )
     return response.data
 }
@@ -147,27 +92,13 @@ export const transferOrganizationOwnership = async (organizationId: string, newO
 // Domain Verification API
 // ============================================================================
 
-export interface OrganizationDomain {
-    id: string
-    slug: string // The actual domain name (e.g., "company.com")
-    name: string | null // Friendly name
-    description: string | null
-    organization_id: string
-    token: string | null // Verification token (available for unverified domains, null for verified)
-    flags: {
-        is_verified?: boolean
-    }
-    created_at: string
-    updated_at: string | null
-}
-
 /**
  * Fetch all domains for an organization
  */
 export const fetchOrganizationDomains = async (): Promise<OrganizationDomain[]> => {
     const response = await axios.get(`${getAgentaApiUrl()}/organizations/domains/`, {
         _ignoreError: true,
-    } as any)
+    })
     return response.data
 }
 
@@ -181,7 +112,7 @@ export const createOrganizationDomain = async (payload: {
 }): Promise<OrganizationDomain> => {
     const response = await axios.post(`${getAgentaApiUrl()}/organizations/domains/`, payload, {
         _ignoreError: true,
-    } as any)
+    })
     return response.data
 }
 
@@ -196,7 +127,7 @@ export const verifyOrganizationDomain = async (domainId: string): Promise<Organi
         },
         {
             _ignoreError: true,
-        } as any,
+        },
     )
     return response.data
 }
@@ -212,7 +143,7 @@ export const refreshOrganizationDomainToken = async (
         undefined,
         {
             _ignoreError: true,
-        } as any,
+        },
     )
     return response.data
 }
@@ -223,39 +154,12 @@ export const refreshOrganizationDomainToken = async (
 export const deleteOrganizationDomain = async (domainId: string): Promise<void> => {
     await axios.delete(`${getAgentaApiUrl()}/organizations/domains/${domainId}`, {
         _ignoreError: true,
-    } as any)
+    })
 }
 
 // ============================================================================
 // SSO/OIDC Provider API
 // ============================================================================
-
-export interface OrganizationProviderSettings {
-    issuer_url?: string
-    client_id?: string
-    client_secret?: string
-    authorization_endpoint?: string
-    token_endpoint?: string
-    userinfo_endpoint?: string
-    scopes?: string[]
-}
-
-export interface OrganizationProvider {
-    id: string
-    slug: string
-    organization_id: string
-    name?: string | null
-    description?: string | null
-    // Backend serves free-form dicts (OrganizationProviderResponse.flags/settings)
-    flags: {
-        is_valid?: boolean
-        is_active?: boolean
-        is_enabled?: boolean
-    }
-    settings: OrganizationProviderSettings
-    created_at: string
-    updated_at: string | null
-}
 
 /**
  * Fetch all SSO providers for an organization
@@ -263,7 +167,7 @@ export interface OrganizationProvider {
 export const fetchOrganizationProviders = async (): Promise<OrganizationProvider[]> => {
     const response = await axios.get(`${getAgentaApiUrl()}/organizations/providers/`, {
         _ignoreError: true,
-    } as any)
+    })
     return response.data
 }
 
@@ -279,7 +183,7 @@ export const createOrganizationProvider = async (payload: {
 }): Promise<OrganizationProvider> => {
     const response = await axios.post(`${getAgentaApiUrl()}/organizations/providers/`, payload, {
         _ignoreError: true,
-    } as any)
+    })
     return response.data
 }
 
@@ -301,7 +205,7 @@ export const updateOrganizationProvider = async (
         payload,
         {
             _ignoreError: true,
-        } as any,
+        },
     )
     return response.data
 }
@@ -317,7 +221,7 @@ export const testOrganizationProvider = async (
         undefined,
         {
             _ignoreError: true,
-        } as any,
+        },
     )
     return response.data
 }
@@ -328,5 +232,5 @@ export const testOrganizationProvider = async (
 export const deleteOrganizationProvider = async (providerId: string): Promise<void> => {
     await axios.delete(`${getAgentaApiUrl()}/organizations/providers/${providerId}`, {
         _ignoreError: true,
-    } as any)
+    })
 }
